Externalized class
